Revert "Migrate to Guice"
Guice is throwing GuiceListener errors.

Run `mvn clean compile jetty:run` to repro.

This reverts commit b1bc837fb3081efef1476267c92e3a425e32076a.

Conflicts:
	src/main/java/com/example/featureswitchservice/Main.java
