+++ src/main/java/com/example/featureswitchservice/FeatureSwitchConfig.java
@@ -2,8 +2,6 @@ package com.example.featureswitchservice;
 
 import com.esotericsoftware.yamlbeans.YamlException;
 
-import javax.inject.Inject;
-import javax.inject.Named;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -13,23 +11,17 @@ import java.io.FileNotFoundException;
 import java.util.Map;
 
 @Path("feature_switch_config")
-public class ConfigResource {
-  @QueryParam("id") String id;
-  @QueryParam("os") String os;
-  @QueryParam("version") String version;
-
-  Map<String, Map<String, String>> config;
-
-  @Inject
-  public ConfigResource(@Named("config") Map<String, Map<String, String>> config) {
-    this.config = config;
-  }
+public class FeatureSwitchConfig {
+    @QueryParam("id") String id;
+    @QueryParam("os") String os;
+    @QueryParam("version") String version;
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public Map<String, Boolean> get() throws FileNotFoundException, YamlException {
       Input input = FeatureSelector.requireInput(new String[]{id, os, version});
-      FeatureSelector selector = new FeatureSelector(this.config);
+      Map<String, Map<String, String>> features = FeatureSelector.createFeatureMap();
+      FeatureSelector selector = new FeatureSelector(features);
       Map<String, Boolean> selected = selector.select(input);
       return selected;
     }
